Refactor the back-shifted Fermi gas model code to use individual objects for each nuclide and to inherit from a generic base class for level density models. Do a bit of code cleanup elsewhere.
Former-commit-id: 84582902d70ded89fa3e57d5b490d0cbedf4469f

+++ src/Integrator.cc
@@ -10,7 +10,7 @@
 double marley::Integrator::num_integrate(const std::function<double(double)> &f,
   double a, double b) const
 {
-  int two_N = 2*N;
+  size_t two_N = 2*N;
   double A = (b - a)/2;
   double B = (b + a)/2;
   double C = (f(b) + f(a))/(two_N);
